extract UTF8StringObjectWriter
